+++ client/fingerprint/consul_test.go
@@ -1,7 +1,8 @@
 package fingerprint
 
 import (
-	"fmt"
+	"io"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -10,207 +11,519 @@ import (
 	"github.com/hashicorp/nomad/client/config"
 	"github.com/hashicorp/nomad/helper/testlog"
 	"github.com/hashicorp/nomad/nomad/structs"
-	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func TestConsulFingerprint(t *testing.T) {
-	fp := NewConsulFingerprint(testlog.HCLogger(t))
-	node := &structs.Node{
-		Attributes: make(map[string]string),
-	}
+// fakeConsul creates an HTTP server mimicking Consul /v1/agent/self endpoint on
+// the first request, and alternates between success and failure responses on
+// subsequent requests
+func fakeConsul(payload string) (*httptest.Server, *config.Config) {
+	working := true
 
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		fmt.Fprintln(w, mockConsulResponse)
-	}))
-	defer ts.Close()
-
-	conf := config.DefaultConfig()
-	conf.ConsulConfig.Addr = strings.TrimPrefix(ts.URL, "http://")
-
-	request := &FingerprintRequest{Config: conf, Node: node}
-	var response FingerprintResponse
-	err := fp.Fingerprint(request, &response)
-	if err != nil {
-		t.Fatalf("Failed to fingerprint: %s", err)
-	}
-
-	if !response.Detected {
-		t.Fatalf("expected response to be applicable")
-	}
-
-	assertNodeAttributeContains(t, response.Attributes, "consul.server")
-	assertNodeAttributeContains(t, response.Attributes, "consul.version")
-	assertNodeAttributeContains(t, response.Attributes, "consul.revision")
-	assertNodeAttributeContains(t, response.Attributes, "unique.consul.name")
-	assertNodeAttributeContains(t, response.Attributes, "consul.datacenter")
-	assertNodeAttributeContains(t, response.Attributes, "consul.segment")
-
-	if _, ok := response.Links["consul"]; !ok {
-		t.Errorf("Expected a link to consul, none found")
-	}
-}
-
-// Taken from tryconsul using consul release 0.5.2
-const mockConsulResponse = `
-{
-  "Config": {
-    "Bootstrap": false,
-    "BootstrapExpect": 3,
-    "Server": true,
-    "Datacenter": "vagrant",
-    "DataDir": "/var/lib/consul",
-    "DNSRecursor": "",
-    "DNSRecursors": [],
-    "DNSConfig": {
-      "NodeTTL": 0,
-      "ServiceTTL": null,
-      "AllowStale": false,
-      "EnableTruncate": false,
-      "MaxStale": 5000000000,
-      "OnlyPassing": false
-    },
-    "Domain": "consul.",
-    "LogLevel": "INFO",
-    "NodeName": "consul2",
-    "ClientAddr": "0.0.0.0",
-    "BindAddr": "0.0.0.0",
-    "AdvertiseAddr": "172.16.59.133",
-    "AdvertiseAddrWan": "172.16.59.133",
-    "Ports": {
-      "DNS": 8600,
-      "HTTP": 8500,
-      "HTTPS": -1,
-      "RPC": 8400,
-      "SerfLan": 8301,
-      "SerfWan": 8302,
-      "Server": 8300
-    },
-    "Addresses": {
-      "DNS": "",
-      "HTTP": "",
-      "HTTPS": "",
-      "RPC": ""
-    },
-    "LeaveOnTerm": false,
-    "SkipLeaveOnInt": false,
-    "StatsiteAddr": "",
-    "StatsitePrefix": "consul",
-    "StatsdAddr": "",
-    "Protocol": 2,
-    "EnableDebug": false,
-    "VerifyIncoming": false,
-    "VerifyOutgoing": false,
-    "VerifyServerHostname": false,
-    "CAFile": "",
-    "CertFile": "",
-    "KeyFile": "",
-    "ServerName": "",
-    "StartJoin": [],
-    "StartJoinWan": [],
-    "RetryJoin": [],
-    "RetryMaxAttempts": 0,
-    "RetryIntervalRaw": "",
-    "RetryJoinWan": [],
-    "RetryMaxAttemptsWan": 0,
-    "RetryIntervalWanRaw": "",
-    "UiDir": "/opt/consul-ui",
-    "PidFile": "",
-    "EnableSyslog": true,
-    "SyslogFacility": "LOCAL0",
-    "RejoinAfterLeave": false,
-    "CheckUpdateInterval": 300000000000,
-    "ACLDatacenter": "",
-    "ACLTTL": 30000000000,
-    "ACLTTLRaw": "",
-    "ACLDefaultPolicy": "allow",
-    "ACLDownPolicy": "extend-cache",
-    "Watches": null,
-    "DisableRemoteExec": false,
-    "DisableUpdateCheck": false,
-    "DisableAnonymousSignature": false,
-    "HTTPAPIResponseHeaders": null,
-    "AtlasInfrastructure": "",
-    "AtlasJoin": false,
-    "Revision": "9a9cc9341bb487651a0399e3fc5e1e8a42e62dd9+CHANGES",
-    "Version": "0.5.2",
-    "VersionPrerelease": "",
-    "UnixSockets": {
-      "Usr": "",
-      "Grp": "",
-      "Perms": ""
-    },
-    "SessionTTLMin": 0,
-    "SessionTTLMinRaw": ""
-  },
-  "Member": {
-    "Name": "consul2",
-    "Addr": "172.16.59.133",
-    "Port": 8301,
-    "Tags": {
-      "build": "0.5.2:9a9cc934",
-      "dc": "vagrant",
-      "expect": "3",
-      "port": "8300",
-      "role": "consul",
-      "segment": "mysegment",
-      "vsn": "2"
-    },
-    "Status": 1,
-    "ProtocolMin": 1,
-    "ProtocolMax": 2,
-    "ProtocolCur": 2,
-    "DelegateMin": 2,
-    "DelegateMax": 4,
-    "DelegateCur": 4
-  }
-}
-`
-
-// TestConsulFingerprint_UnexpectedResponse asserts that the Consul
-// fingerprinter does not panic when it encounters an unexpected response.
-// See https://github.com/hashicorp/nomad/issues/3326
-func TestConsulFingerprint_UnexpectedResponse(t *testing.T) {
-	assert := assert.New(t)
-	fp := NewConsulFingerprint(testlog.HCLogger(t))
-	node := &structs.Node{
-		Attributes: make(map[string]string),
-	}
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		fmt.Fprintln(w, "{}")
-	}))
-	defer ts.Close()
-
-	conf := config.DefaultConfig()
-	conf.ConsulConfig.Addr = strings.TrimPrefix(ts.URL, "http://")
-
-	request := &FingerprintRequest{Config: conf, Node: node}
-	var response FingerprintResponse
-	err := fp.Fingerprint(request, &response)
-	assert.Nil(err)
-
-	if !response.Detected {
-		t.Fatalf("expected response to be applicable")
-	}
-
-	attrs := []string{
-		"consul.server",
-		"consul.version",
-		"consul.revision",
-		"unique.consul.name",
-		"consul.datacenter",
-		"consul.segment",
-	}
-
-	for _, attr := range attrs {
-		if v, ok := response.Attributes[attr]; ok {
-			t.Errorf("unexpected node attribute %q with vlaue %q", attr, v)
+		if working {
+			_, _ = io.WriteString(w, payload)
+			working = false
+		} else {
+			w.WriteHeader(http.StatusInternalServerError)
+			working = true
 		}
-	}
+	}))
 
-	if v, ok := response.Links["consul"]; ok {
-		t.Errorf("Unexpected link to consul: %v", v)
-	}
+	cfg := config.DefaultConfig()
+	cfg.ConsulConfig.Addr = strings.TrimPrefix(ts.URL, `http://`)
+	return ts, cfg
+}
+
+func fakeConsulPayload(t *testing.T, filename string) string {
+	b, err := ioutil.ReadFile(filename)
+	require.NoError(t, err)
+	return string(b)
+}
+
+func newConsulFingerPrint(t *testing.T) *ConsulFingerprint {
+	return NewConsulFingerprint(testlog.HCLogger(t)).(*ConsulFingerprint)
+}
+
+func TestConsulFingerprint_server(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("is server", func(t *testing.T) {
+		s, ok := fp.server(consulInfo{
+			"Config": {"Server": true},
+		})
+		require.True(t, ok)
+		require.Equal(t, "true", s)
+	})
+
+	t.Run("is not server", func(t *testing.T) {
+		s, ok := fp.server(consulInfo{
+			"Config": {"Server": false},
+		})
+		require.True(t, ok)
+		require.Equal(t, "false", s)
+	})
+
+	t.Run("missing", func(t *testing.T) {
+		_, ok := fp.server(consulInfo{
+			"Config": {},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("malformed", func(t *testing.T) {
+		_, ok := fp.server(consulInfo{
+			"Config": {"Server": 9000},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_version(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("oss", func(t *testing.T) {
+		v, ok := fp.version(consulInfo{
+			"Config": {"Version": "v1.9.5"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "v1.9.5", v)
+	})
+
+	t.Run("ent", func(t *testing.T) {
+		v, ok := fp.version(consulInfo{
+			"Config": {"Version": "v1.9.5+ent"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "v1.9.5+ent", v)
+	})
+
+	t.Run("missing", func(t *testing.T) {
+		_, ok := fp.version(consulInfo{
+			"Config": {},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("malformed", func(t *testing.T) {
+		_, ok := fp.version(consulInfo{
+			"Config": {"Version": 9000},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_sku(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("oss", func(t *testing.T) {
+		s, ok := fp.sku(consulInfo{
+			"Config": {"Version": "v1.9.5"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "oss", s)
+	})
+
+	t.Run("oss dev", func(t *testing.T) {
+		s, ok := fp.sku(consulInfo{
+			"Config": {"Version": "v1.9.5-dev"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "oss", s)
+	})
+
+	t.Run("ent", func(t *testing.T) {
+		s, ok := fp.sku(consulInfo{
+			"Config": {"Version": "v1.9.5+ent"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "ent", s)
+	})
+
+	t.Run("ent dev", func(t *testing.T) {
+		s, ok := fp.sku(consulInfo{
+			"Config": {"Version": "v1.9.5+ent-dev"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "ent", s)
+	})
+
+	t.Run("missing", func(t *testing.T) {
+		_, ok := fp.sku(consulInfo{
+			"Config": {},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("malformed", func(t *testing.T) {
+		_, ok := fp.sku(consulInfo{
+			"Config": {"Version": "***"},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_revision(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("ok", func(t *testing.T) {
+		r, ok := fp.revision(consulInfo{
+			"Config": {"Revision": "3c1c22679"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "3c1c22679", r)
+	})
+
+	t.Run("malformed", func(t *testing.T) {
+		_, ok := fp.revision(consulInfo{
+			"Config": {"Revision": 9000},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("missing", func(t *testing.T) {
+		_, ok := fp.revision(consulInfo{
+			"Config": {},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_dc(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("ok", func(t *testing.T) {
+		dc, ok := fp.dc(consulInfo{
+			"Config": {"Datacenter": "dc1"},
+		})
+		require.True(t, ok)
+		require.Equal(t, "dc1", dc)
+	})
+
+	t.Run("malformed", func(t *testing.T) {
+		_, ok := fp.dc(consulInfo{
+			"Config": {"Datacenter": 9000},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("missing", func(t *testing.T) {
+		_, ok := fp.dc(consulInfo{
+			"Config": {},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_segment(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("ok", func(t *testing.T) {
+		s, ok := fp.segment(consulInfo{
+			"Member": {"Tags": map[string]interface{}{"segment": "seg1"}},
+		})
+		require.True(t, ok)
+		require.Equal(t, "seg1", s)
+	})
+
+	t.Run("segment missing", func(t *testing.T) {
+		_, ok := fp.segment(consulInfo{
+			"Member": {"Tags": map[string]interface{}{}},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("tags missing", func(t *testing.T) {
+		_, ok := fp.segment(consulInfo{
+			"Member": {},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("malformed", func(t *testing.T) {
+		_, ok := fp.segment(consulInfo{
+			"Member": {"Tags": map[string]interface{}{"segment": 9000}},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_connect(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("connect enabled", func(t *testing.T) {
+		s, ok := fp.connect(consulInfo{
+			"DebugConfig": {"ConnectEnabled": true},
+		})
+		require.True(t, ok)
+		require.Equal(t, "true", s)
+	})
+
+	t.Run("connect not enabled", func(t *testing.T) {
+		s, ok := fp.connect(consulInfo{
+			"DebugConfig": {"ConnectEnabled": false},
+		})
+		require.True(t, ok)
+		require.Equal(t, "false", s)
+	})
+
+	t.Run("connect missing", func(t *testing.T) {
+		_, ok := fp.connect(consulInfo{
+			"DebugConfig": {},
+		})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_grpc(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("grpc set", func(t *testing.T) {
+		s, ok := fp.grpc(consulInfo{
+			"DebugConfig": {"GRPCPort": 8502.0}, // JSON numbers are floats
+		})
+		require.True(t, ok)
+		require.Equal(t, "8502", s)
+	})
+
+	t.Run("grpc disabled", func(t *testing.T) {
+		s, ok := fp.grpc(consulInfo{
+			"DebugConfig": {"GRPCPort": -1.0}, // JSON numbers are floats
+		})
+		require.True(t, ok)
+		require.Equal(t, "-1", s)
+	})
+
+	t.Run("grpc missing", func(t *testing.T) {
+		_, ok := fp.grpc(consulInfo{
+			"DebugConfig": {},
+		})
+		require.False(t, ok)
+	})
+
+}
+
+func TestConsulFingerprint_namespaces(t *testing.T) {
+	t.Parallel()
+
+	fp := newConsulFingerPrint(t)
+
+	t.Run("supports namespaces", func(t *testing.T) {
+		s, ok := fp.namespaces(consulInfo{
+			"Stats": {"license": map[string]interface{}{"features": "Automated Backups, Automated Upgrades, Enhanced Read Scalability, Network Segments, Redundancy Zone, Advanced Network Federation, Namespaces, SSO, Audit Logging"}},
+		})
+		require.True(t, ok)
+		require.Equal(t, "true", s)
+	})
+
+	t.Run("no namespaces", func(t *testing.T) {
+		_, ok := fp.namespaces(consulInfo{
+			"Stats": {"license": map[string]interface{}{"features": "Automated Backups, Automated Upgrades, Enhanced Read Scalability, Network Segments, Redundancy Zone, Advanced Network Federation, SSO, Audit Logging"}},
+		})
+		require.False(t, ok)
+	})
+
+	t.Run("stats missing", func(t *testing.T) {
+		_, ok := fp.namespaces(consulInfo{})
+		require.False(t, ok)
+	})
+
+	t.Run("license missing", func(t *testing.T) {
+		_, ok := fp.namespaces(consulInfo{"Stats": {}})
+		require.False(t, ok)
+	})
+
+	t.Run("features missing", func(t *testing.T) {
+		_, ok := fp.namespaces(consulInfo{"Stats": {"license": map[string]interface{}{}}})
+		require.False(t, ok)
+	})
+}
+
+func TestConsulFingerprint_Fingerprint_oss(t *testing.T) {
+	cf := newConsulFingerPrint(t)
+
+	ts, cfg := fakeConsul(fakeConsulPayload(t, "test_fixtures/consul/agent_self_oss.json"))
+	defer ts.Close()
+
+	node := &structs.Node{Attributes: make(map[string]string)}
+
+	// consul not available before first run
+	require.Equal(t, consulUnavailable, cf.lastState)
+
+	// execute first query with good response
+	var resp FingerprintResponse
+	err := cf.Fingerprint(&FingerprintRequest{Config: cfg, Node: node}, &resp)
+	require.NoError(t, err)
+	require.Equal(t, map[string]string{
+		"consul.datacenter":  "dc1",
+		"consul.revision":    "3c1c22679",
+		"consul.segment":     "seg1",
+		"consul.server":      "true",
+		"consul.sku":         "oss",
+		"consul.version":     "1.9.5",
+		"consul.connect":     "true",
+		"consul.grpc":        "8502",
+		"unique.consul.name": "HAL9000",
+	}, resp.Attributes)
+	require.True(t, resp.Detected)
+
+	// consul now available
+	require.Equal(t, consulAvailable, cf.lastState)
+
+	var resp2 FingerprintResponse
+
+	// pretend attributes set for failing request
+	node.Attributes["consul.datacenter"] = "foo"
+	node.Attributes["consul.revision"] = "foo"
+	node.Attributes["consul.segment"] = "foo"
+	node.Attributes["consul.server"] = "foo"
+	node.Attributes["consul.sku"] = "foo"
+	node.Attributes["consul.version"] = "foo"
+	node.Attributes["consul.connect"] = "foo"
+	node.Attributes["connect.grpc"] = "foo"
+	node.Attributes["unique.consul.name"] = "foo"
+
+	// execute second query with error
+	err2 := cf.Fingerprint(&FingerprintRequest{Config: cfg, Node: node}, &resp2)
+	require.NoError(t, err2)            // does not return error
+	require.Equal(t, map[string]string{ // attributes set empty
+		"consul.datacenter":    "",
+		"consul.revision":      "",
+		"consul.segment":       "",
+		"consul.server":        "",
+		"consul.sku":           "",
+		"consul.version":       "",
+		"unique.consul.name":   "",
+		"consul.connect":       "",
+		"consul.grpc":          "",
+		"consul.ft.namespaces": "",
+	}, resp2.Attributes)
+	require.True(t, resp.Detected) // never downgrade
+
+	// consul no longer available
+	require.Equal(t, consulUnavailable, cf.lastState)
+
+	// execute third query no error
+	var resp3 FingerprintResponse
+	err3 := cf.Fingerprint(&FingerprintRequest{Config: cfg, Node: node}, &resp3)
+	require.NoError(t, err3)
+	require.Equal(t, map[string]string{
+		"consul.datacenter":  "dc1",
+		"consul.revision":    "3c1c22679",
+		"consul.segment":     "seg1",
+		"consul.server":      "true",
+		"consul.sku":         "oss",
+		"consul.version":     "1.9.5",
+		"consul.connect":     "true",
+		"consul.grpc":        "8502",
+		"unique.consul.name": "HAL9000",
+	}, resp3.Attributes)
+
+	// consul now available again
+	require.Equal(t, consulAvailable, cf.lastState)
+	require.True(t, resp.Detected)
+}
+
+func TestConsulFingerprint_Fingerprint_ent(t *testing.T) {
+	cf := newConsulFingerPrint(t)
+
+	ts, cfg := fakeConsul(fakeConsulPayload(t, "test_fixtures/consul/agent_self_ent.json"))
+	defer ts.Close()
+
+	node := &structs.Node{Attributes: make(map[string]string)}
+
+	// consul not available before first run
+	require.Equal(t, consulUnavailable, cf.lastState)
+
+	// execute first query with good response
+	var resp FingerprintResponse
+	err := cf.Fingerprint(&FingerprintRequest{Config: cfg, Node: node}, &resp)
+	require.NoError(t, err)
+	require.Equal(t, map[string]string{
+		"consul.datacenter":    "dc1",
+		"consul.revision":      "22ce6c6ad",
+		"consul.segment":       "seg1",
+		"consul.server":        "true",
+		"consul.sku":           "ent",
+		"consul.version":       "1.9.5+ent",
+		"consul.ft.namespaces": "true",
+		"consul.connect":       "true",
+		"consul.grpc":          "8502",
+		"unique.consul.name":   "HAL9000",
+	}, resp.Attributes)
+	require.True(t, resp.Detected)
+
+	// consul now available
+	require.Equal(t, consulAvailable, cf.lastState)
+
+	var resp2 FingerprintResponse
+
+	// pretend attributes set for failing request
+	node.Attributes["consul.datacenter"] = "foo"
+	node.Attributes["consul.revision"] = "foo"
+	node.Attributes["consul.segment"] = "foo"
+	node.Attributes["consul.server"] = "foo"
+	node.Attributes["consul.sku"] = "foo"
+	node.Attributes["consul.version"] = "foo"
+	node.Attributes["consul.ft.namespaces"] = "foo"
+	node.Attributes["consul.connect"] = "foo"
+	node.Attributes["connect.grpc"] = "foo"
+	node.Attributes["unique.consul.name"] = "foo"
+
+	// execute second query with error
+	err2 := cf.Fingerprint(&FingerprintRequest{Config: cfg, Node: node}, &resp2)
+	require.NoError(t, err2)            // does not return error
+	require.Equal(t, map[string]string{ // attributes set empty
+		"consul.datacenter":    "",
+		"consul.revision":      "",
+		"consul.segment":       "",
+		"consul.server":        "",
+		"consul.sku":           "",
+		"consul.version":       "",
+		"consul.ft.namespaces": "",
+		"consul.connect":       "",
+		"consul.grpc":          "",
+		"unique.consul.name":   "",
+	}, resp2.Attributes)
+	require.True(t, resp.Detected) // never downgrade
+
+	// consul no longer available
+	require.Equal(t, consulUnavailable, cf.lastState)
+
+	// execute third query no error
+	var resp3 FingerprintResponse
+	err3 := cf.Fingerprint(&FingerprintRequest{Config: cfg, Node: node}, &resp3)
+	require.NoError(t, err3)
+	require.Equal(t, map[string]string{
+		"consul.datacenter":    "dc1",
+		"consul.revision":      "22ce6c6ad",
+		"consul.segment":       "seg1",
+		"consul.server":        "true",
+		"consul.sku":           "ent",
+		"consul.version":       "1.9.5+ent",
+		"consul.ft.namespaces": "true",
+		"consul.connect":       "true",
+		"consul.grpc":          "8502",
+		"unique.consul.name":   "HAL9000",
+	}, resp3.Attributes)
+
+	// consul now available again
+	require.Equal(t, consulAvailable, cf.lastState)
+	require.True(t, resp.Detected)
 }